+++ contracts/interfaces/ICollateralReceiver.sol
@@ -3,5 +3,8 @@
 pragma solidity 0.8.6;
 
 interface ICollateralReceiver {
-    function receiveCollateral(address[] memory _tokens, uint[] memory _amounts) external;
+    function receiveCollateral(
+        address[] memory _tokens,
+        uint256[] memory _amounts
+    ) external;
 }
