+++ contracts/interfaces/IPool.sol
@@ -6,30 +6,34 @@ import "./ICollateralReceiver.sol";
 
 // Common interface for the Pools.
 interface IPool is ICollateralReceiver {
-    
     // --- Events ---
-    
-    event ETHBalanceUpdated(uint _newBalance);
-    event YUSDBalanceUpdated(uint _newBalance);
+
+    event ETHBalanceUpdated(uint256 _newBalance);
+    event YUSDBalanceUpdated(uint256 _newBalance);
     event ActivePoolAddressChanged(address _newActivePoolAddress);
     event DefaultPoolAddressChanged(address _newDefaultPoolAddress);
     event StabilityPoolAddressChanged(address _newStabilityPoolAddress);
     event WhitelistAddressChanged(address _newWhitelistAddress);
-    event EtherSent(address _to, uint _amount);
-    event CollateralSent(address _collateral, address _to, uint _amount);
+    event EtherSent(address _to, uint256 _amount);
+    event CollateralSent(address _collateral, address _to, uint256 _amount);
 
     // --- Functions ---
 
-    function getVC() external view returns (uint);
+    function getVC() external view returns (uint256);
 
-    function getCollateral(address collateralAddress) external view returns (uint);
+    function getCollateral(address collateralAddress)
+        external
+        view
+        returns (uint256);
 
-    function getAllCollateral() external view returns (address[] memory, uint256[] memory);
+    function getAllCollateral()
+        external
+        view
+        returns (address[] memory, uint256[] memory);
 
-    function getYUSDDebt() external view returns (uint);
+    function getYUSDDebt() external view returns (uint256);
 
-    function increaseYUSDDebt(uint _amount) external;
+    function increaseYUSDDebt(uint256 _amount) external;
 
-    function decreaseYUSDDebt(uint _amount) external;
-
+    function decreaseYUSDDebt(uint256 _amount) external;
 }
